# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HawaiiConditions",
    "version": "1.0.5",
    "description": "HawaiiConditions — real-time Hawaii conditions with prepaid-balance payments via Stripe card."
  },
  "paths": {
    "/api/mcp": {
      "get": {
        "operationId": "mcp_info",
        "summary": "MCP server info and tool list",
        "tags": [
          "MCP"
        ],
        "x-payment-info": {
          "price": 0,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "free"
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Server info and tools"
          }
        }
      },
      "post": {
        "operationId": "mcp_message",
        "summary": "Send MCP JSON-RPC 2.0 message",
        "tags": [
          "MCP"
        ],
        "x-payment-info": {
          "price": 0,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "free"
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Message processed"
          },
          "402": {
            "description": "Payment required"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string"
                  },
                  "method": {
                    "type": "string"
                  }
                },
                "required": [
                  "jsonrpc",
                  "method"
                ]
              }
            }
          }
        }
      }
    },
    "/gpt/weather": {
      "get": {
        "operationId": "get_weather",
        "summary": "Get current Hawaii weather conditions",
        "tags": [
          "Hawaii Conditions"
        ],
        "x-payment-info": {
          "price": 0.1,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "prepaid_balance",
          "accountHeader": "X-MCP-Account"
        },
        "parameters": [
          {
            "name": "X-MCP-Account",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^mcp_live_"
            },
            "description": "Your prepaid API key (format: mcp_live_…). Register via register_agent."
          },
          {
            "name": "island",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "oahu",
                "maui",
                "kauai",
                "big-island",
                "molokai",
                "lanai"
              ],
              "default": "oahu"
            }
          }
        ],
        "security": [
          {
            "McpPrepaidAccount": []
          },
          {
            "MppPayment": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "Payment required"
          }
        }
      },
      "post": {
        "operationId": "post_weather",
        "summary": "Get current Hawaii weather conditions (POST)",
        "tags": [
          "Hawaii Conditions"
        ],
        "x-payment-info": {
          "price": 0.1,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "prepaid_balance",
          "accountHeader": "X-MCP-Account"
        },
        "parameters": [
          {
            "name": "X-MCP-Account",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^mcp_live_"
            },
            "description": "Your prepaid API key (format: mcp_live_…). Register via register_agent."
          }
        ],
        "security": [
          {
            "McpPrepaidAccount": []
          },
          {
            "MppPayment": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "Payment required"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "island": {
                    "type": "string",
                    "enum": [
                      "oahu",
                      "maui",
                      "kauai",
                      "big-island",
                      "molokai",
                      "lanai"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/gpt/ocean-safety": {
      "get": {
        "operationId": "get_ocean_safety",
        "summary": "Get Hawaii ocean safety conditions",
        "tags": [
          "Hawaii Conditions"
        ],
        "x-payment-info": {
          "price": 0.5,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "prepaid_balance",
          "accountHeader": "X-MCP-Account"
        },
        "parameters": [
          {
            "name": "X-MCP-Account",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^mcp_live_"
            },
            "description": "Your prepaid API key (format: mcp_live_…). Register via register_agent."
          },
          {
            "name": "island",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "oahu",
                "maui",
                "kauai",
                "big-island",
                "molokai",
                "lanai"
              ],
              "default": "oahu"
            }
          }
        ],
        "security": [
          {
            "McpPrepaidAccount": []
          },
          {
            "MppPayment": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "Payment required"
          }
        }
      },
      "post": {
        "operationId": "post_ocean_safety",
        "summary": "Get Hawaii ocean safety conditions (POST)",
        "tags": [
          "Hawaii Conditions"
        ],
        "x-payment-info": {
          "price": 0.5,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "prepaid_balance",
          "accountHeader": "X-MCP-Account"
        },
        "parameters": [
          {
            "name": "X-MCP-Account",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^mcp_live_"
            },
            "description": "Your prepaid API key (format: mcp_live_…). Register via register_agent."
          }
        ],
        "security": [
          {
            "McpPrepaidAccount": []
          },
          {
            "MppPayment": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "Payment required"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "island": {
                    "type": "string",
                    "enum": [
                      "oahu",
                      "maui",
                      "kauai",
                      "big-island",
                      "molokai",
                      "lanai"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gpt/sun-times": {
      "get": {
        "operationId": "get_sun_times",
        "summary": "Get Hawaii sunrise and sunset times — Free",
        "tags": [
          "Hawaii Conditions"
        ],
        "x-payment-info": {
          "price": 0,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "free"
        },
        "parameters": [
          {
            "name": "X-MCP-Account",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^mcp_live_"
            },
            "description": "Your prepaid API key (format: mcp_live_…). Register via register_agent."
          },
          {
            "name": "island",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "oahu",
                "maui",
                "kauai",
                "big-island",
                "molokai",
                "lanai"
              ],
              "default": "oahu"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "operationId": "post_sun_times",
        "summary": "Get Hawaii sunrise and sunset times (POST) — Free",
        "tags": [
          "Hawaii Conditions"
        ],
        "x-payment-info": {
          "price": 0,
          "currency": "USD",
          "protocols": [
            "stripe-card-prepaid"
          ],
          "protocol": "stripe-card-prepaid",
          "pricingMode": "fixed",
          "billing": "free"
        },
        "parameters": [
          {
            "name": "X-MCP-Account",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^mcp_live_"
            },
            "description": "Your prepaid API key (format: mcp_live_…). Register via register_agent."
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "island": {
                    "type": "string",
                    "enum": [
                      "oahu",
                      "maui",
                      "kauai",
                      "big-island",
                      "molokai",
                      "lanai"
                    ]
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "McpPrepaidAccount": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MCP-Account",
        "description": "Prepaid account key (format: mcp_live_…). Register via register_agent at https://hawaii-conditions-prime.vercel.app/mcp"
      },
      "MppPayment": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "MPP",
        "description": "MPP payment credential issued after a 402 challenge."
      }
    }
  }
}